refactor pie responsive options to pull from chart options helper file
