+++ src/components/ActiveJob.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import Chartist from 'chartist';
 import ChartistGraph from 'react-chartist';
+import pieResponsiveOptions from '../helpers/ChartOptions';
+
 
 class ActiveJob extends React.Component {
   render() {
@@ -26,16 +28,6 @@ class ActiveJob extends React.Component {
       ]
     };
 
-    const pieResponsiveOptions = [
-      ['screen and (max-width: 576px)', {
-        height: '200px',
-        donutWidth: 10,
-      }],
-      ['screen and (min-width: 577px) and (max-width: 768px)', {
-        height: '250px',
-        donutWidth: 60,
-      }]
-    ];
 
     const data = {
       labels: [
