+++ src/components/RailsUpdates.js
@@ -2,6 +2,8 @@ import React from 'react';
 import Chartist from 'chartist';
 import ChartistGraph from 'react-chartist';
 import ChartistPluginLegend from 'chartist-plugin-legend';
+import pieResponsiveOptions from '../helpers/ChartOptions';
+
 
 class RailsUpdates extends React.Component {
   render() {
@@ -22,17 +24,6 @@ class RailsUpdates extends React.Component {
         }),
       ]
     };
-
-    const pieResponsiveOptions = [
-      ['screen and (max-width: 576px)', {
-        height: '200px',
-        donutWidth: 50,
-      }],
-      ['screen and (min-width: 577px) and (max-width: 768px)', {
-        height: '250px',
-        donutWidth: 60,
-      }]
-    ];
 
     const releasesPieChartData = {
       labels: [
